+++ src/azure-cli/azure/cli/command_modules/storage/_params.py
@@ -240,6 +240,14 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('vnet_name', help='Name of a virtual network.', validator=validate_subnet)
         c.argument('action', help='The action of virtual network rule.')
 
+    with self.argument_context('storage account blob-service-properties show') as c:
+        c.argument('account_name', acct_name_type, id_part=None)
+
+    with self.argument_context('storage account blob-service-properties update') as c:
+        c.argument('account_name', acct_name_type, id_part=None)
+        c.argument('enable_change_feed', arg_type=get_three_state_flag(),
+                   help='Indicates whether change feed event logging is enabled.')
+
     with self.argument_context('storage account generate-sas') as c:
         t_account_permissions = self.get_sdk('common.models#AccountPermissions')
         c.register_sas_arguments()