+++ src/azure-cli/HISTORY.rst
@@ -23,13 +23,14 @@ Release History
 
 * GA Release Large File Shares property for storage account create and update command
 * GA Release User Delegation SAS token Support
+* Add new commands to manage blob service properties for storage account.
 
 2.0.77
 ++++++
 
 **ACR**
 
-* Deprecated paramater `--branch` from acr task create/update
+* Deprecated parameter `--branch` from acr task create/update
 
 **Azure Red Hat OpenShift**